version 1.1.11 debug and add continue btn event
add event to addContinueBtnEvent() which targets the continue btn, when
clicked it will update the current players grid with all the current
"filled" elements on the display.

debug dragLeaveEvent in dragdropevent.js, bug removes the 'data-ship'
attribute on an "occupied" grid elem, breaking the logic behind
interacting the game object.