version 1.4.0 add gameevent.js
add gameevent.js that will contain logic when a grid element is clicked
on a player's grid board.

debug setPlayerGrid method in accesoryfunc.js, which sets all the ship
obj on the players grid into different individual ship objects instead
of referring to a singular ship obj. refactored setPlayergrid method
moving the shipObjArr variable on top of the code block to prevent
looping and creating new array of ship obj