version 1.4.1 refactor gameevent.js
playerOneGridElemEvent and playerTwoGridElemEvent are event functions
that will be added to each respective grid elements. Both function do
similar things, clicking the elements will interact with grid obj
invoking the receiveAttack method, if it's a hit, it will also check
if the ship sunk, else it will just switch the game state to the other
players turn. At the end of the event it will scan for the board if
the board still contains ship.