version 2.0.0 feat: playable versus computer
add placeRandomShip method in GameBoard class in gameboard.js. This
method is invoked when playing against the computer.

add several methods in Player class that the computer opponent will
use against the user. playMove method contains the computer logic
where it generates a random number and use it to hit the user's grid,
it also contains a logic where it checks if the random move has been
used already.

refactor playerTwoGridElemEvent() in gamevent.js into async func.
When the user click the computer's grid and missed, it will switch to
computers turn which is a while loop. Inside the loop it will wait for
the compAttack() function to return a promise. if returned true it will
continuously attack, if false it will break the loop and switch to
player ones turn.