version 2.0.3 debug grid elems events
debug playerOneGridElemEVent and playerTwoGridElemEVent. after ending
the first game, clicking on a grid elem will trigger the event to scan
the grid, which will scan the empty grid thus invoking the endgame
function.

refactor playerOneGridElemEVent and playerTwoGridElemEVent, by enclosing
the while code block into an if statement based on the state of the
game obj.